feat: incluir horarios y agencias de paradas hija - limitar resultados horarios a 2 horas o lo que se pida

+++ api/models/viaje.py
@@ -14,7 +14,7 @@ class HorariosViajeModel(BaseModel):
   tipoBajada: int | None = None
   recogidaContinua: bool | None = None
   bajadaContinua: bool | None = None
-  distanciaRecorrida: int | None = None
+  distanciaRecorrida: float | None = None
   exacto: bool | None = None
 
 class FrecuenciasViajeModel(BaseModel):
